fix(im): 站内信/IM安全加固 + 推荐系统优化 + 审核责任链/LLM结构化输出

+++ heima-leadnews/heima-leadnews-service/heima-leadnews-content/src/main/java/com/heima/content/service/article/impl/ArticleSimilarityServiceImpl.java
@@ -40,7 +40,7 @@ public class ArticleSimilarityServiceImpl implements ArticleSimilarityService {
             if (similarArticles != null && !similarArticles.isEmpty()) {
                 // 排除自身
                 Object[] mostSimilar = similarArticles.stream()
-                        .filter(arr -> !((Long) arr[0]).equals(article.getId()))
+                        .filter(arr -> !arr[0].equals(article.getId()))
                         .findFirst()
                         .orElse(null);
